+++ htdocs/ServiceDeskSpecialAttentionAlert.php
@@ -7,10 +7,11 @@
  * @authors Karim Ahmed - Sweet Code Limited
  */
 require_once("config.inc.php");
-GLOBAL $cfg;
+global $cfg;
 require_once($cfg['path_bu'] . '/BUProblemSLA.inc.php');
 $thing = null;
 $buProblemSLA = new BUProblemSLA($thing);
 $buProblemSLA->specialAttentionEmailAlert();
+$buProblemSLA->clearBreachedSpecialAttentionCustomers();
 echo "Special Attention Customer Email Alert Routine Finished";
 ?>