+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project changes will be shown here.
 - Issue 961 - Team Performance at the start of the month
 ### Changed
 - Issue 964 - Make Office 365 Backup unit numbers field read only because it's linked to an API
+### Fixed
+- Issue 968 - Clear Special Attention Status Once Expired
 
 ## [v4.3.4] - 2020-08-10
 ### Fixed